+++ packages/assignment-7/src/App.tsx
@@ -37,15 +37,10 @@ import {
 } from '@chakra-ui/icons';
 import useEvents from './hooks/useEvents';
 import { NOTIFICATION_OPTIONS_DUMMY, WEEK_DAYS_DUMMY } from './dummys';
-import {
-  formatMonth,
-  formatWeek,
-  getDaysInMonth,
-  getWeekDates,
-  navigate,
-} from './utils';
+import { formatMonth, getDaysInMonth, navigate } from './utils';
 import UseFetchHolidays from './hooks/useFetchHolidays';
 import EventForm from './components/EventForm';
+import WeekView from './components/WeekView';
 
 function App() {
   const {
@@ -99,63 +94,12 @@ function App() {
   const holidays = UseFetchHolidays(currentDate);
 
   const renderWeekView = () => {
-    const weekDates = getWeekDates(currentDate);
     return (
-      <VStack data-testid="week-view" align="stretch" w="full" spacing={4}>
-        <Heading size="md">{formatWeek(currentDate)}</Heading>
-        <Table variant="simple" w="full">
-          <Thead>
-            <Tr>
-              {WEEK_DAYS_DUMMY.map((day) => (
-                <Th key={day} width="14.28%">
-                  {day}
-                </Th>
-              ))}
-            </Tr>
-          </Thead>
-          <Tbody>
-            <Tr>
-              {weekDates.map((date) => (
-                <Td
-                  key={date.toISOString()}
-                  height="100px"
-                  verticalAlign="top"
-                  width="14.28%"
-                >
-                  <Text fontWeight="bold">{date.getDate()}</Text>
-                  {filteredEvents
-                    .filter(
-                      (event) =>
-                        new Date(event.date).toDateString() ===
-                        date.toDateString()
-                    )
-                    .map((event) => {
-                      const isNotified = notifiedEvents.includes(event.id);
-                      return (
-                        <Box
-                          key={event.id}
-                          p={1}
-                          my={1}
-                          bg={isNotified ? 'red.100' : 'gray.100'}
-                          borderRadius="md"
-                          fontWeight={isNotified ? 'bold' : 'normal'}
-                          color={isNotified ? 'red.500' : 'inherit'}
-                        >
-                          <HStack spacing={1}>
-                            {isNotified && <BellIcon />}
-                            <Text fontSize="sm" noOfLines={1}>
-                              {event.title}
-                            </Text>
-                          </HStack>
-                        </Box>
-                      );
-                    })}
-                </Td>
-              ))}
-            </Tr>
-          </Tbody>
-        </Table>
-      </VStack>
+      <WeekView
+        currentDate={currentDate}
+        filteredEvents={filteredEvents}
+        notifiedEvents={notifiedEvents}
+      />
     );
   };
